Noticed issues with Autohotkey (AHK).
I map some keys (F9 and F11 to input date/datetime) to use this
in all applications.  This causes weird problems with normal mode in
vim.  If you observe strange key behvaior in normal mode and you cannot
locate keymaping in your vimrc file try to verify that the key binding
is not coming from AHK.