Added some more options for psql.
Learned about the versatility of psql from Dimitri Fontaine's "The Art
of PostgreSQL". He mentioned in the book that the default `psql` client
that comes wit PostgreSQL is quite powerful and suggested reading the
manual page thoroughly. So, I did and was glad to learn many of the
features present in `psql`.

Setting ECHO_HIDDEN 'on' will display the underlying query used by many
of the meta commands. This is a good way to get an understanding of
PostgreSQL's internals.