Use vim for reading manpages.
This enables :Man command in vim. This is very helpful because I can use
all the vim commands to navigate the manpage with the added advantage
that all the related manpages become tags that can be navigated quite
easily. What does that mean? It means that typing C-] on any of the "See
Also" manpages will actually take me to the linked manpage and C-t will
bring me back to the previous page. I am annoyed that I did not know of
this earlier!