Don't add VROOT (virtual root) to path env variable multiple times.
Nix development shell opens a bash shell which runs the ~/.bashrc file
during initialization. Calling a new development shell inside another
development shell kept on appending VROOT env variable to PATH. To avoid
this check if the variable is zero length string and if so append it to
the path. If the variable is not a zero length string then it is already
set and there is nothing to be done.

ZSH conveniently provides `typeset -U` (see `man zshbuiltins`) which
removes duplicates from an array. And, since path and PATH variables are
different representations of the same thing we can just use typeset -U
to remove duplicates. There is nothing like this in Bash.

I think this also makes the initialization script more idempotent.