Use fully qualified table name.
Dimitri Fontaine's sql code from "The Art of PostgreSQL" (TAOP) is
organized in schema and tables. Loading the code accompanied with the
book will create schemas that are not included in the search_path. So,
when you do the `\d` command the tables in the schemas not in the
search_path don't show up. But the system view `pg_tables` lists all the
tables in all the schemas! So, to get the size, or number of rows, of
table the name needs to be qualified with the schemaname.