Modified the custom command to show more details for columns.
After using the `:cols` custom command for a bit I realized that
I also needed to know the type of the column. Also, using `\x auto` was
causing wide tables (especially GIS imported data) to switch to using
extended display mode which is not conducive for interactive use.

An additional benefit of this is that after running `:cols` you can run
the `\e` (edit meta-command) and modify the `where` clause to narrow
down your query.