Removing `-r` broke terminal coloring...especially for git commands.
The reason I removed `-r` from LESS envvar was that it messes up comment
lines, especially at the start of the file! Typing `less my-file.py` on
the terminal would often not show the top few lines of the comments.
I don't know if this is because of LESS or LESSOPEN envvar settings.
I do not have LESSOPEN set so I doubt that that is causing this issue.