Show shcemaname along with tablename.
Even though PostgreSQL correctly displays schema-qualified names on
output if the object is not found in the current search_path it is
a good idea to include schemaname in the query output. The indexdef
correctly shows the schema qualified name in the index definition but
not including the schemaname in the output result does not highlight
that the definition is not visible from the current search_path. For
isntance, `\di` won't show the indexes which will appear in the `:indexes`
command. And, that's because `:indexes` is looking for all indexes
whether they're in the schema in the search_path or not.