Consider access control permissions for pg_database_size.
ACL are often set in multiuser databases which can cause this query to
fail with permission issues. So, if the current_user does not have
appropriate permissions don't print the relevant information instead of
failing with permission errors! psql's meta commands also use these
permission checks. Check out \l+ when psql is called with -E (or psql
variable ECHO_HIDDEN is set to on) to see how this privilege check is
used.