Command to set search_path in psql.
This command is especially needed when connecting to a different
database using the \c command in psql. This newly connected database will have
different schemas which will, most likely, not be in the search_path
which makes the other psql meta commands less useful. This command
simplifies setting the search_path to all accessible schemas. By the
way, this is also done for the first database that I connect to!