Escape underscore character to enable explicit match in like clause
An underscore in pattern stands for a any single character; a percent
(%) matches any sequence of zero or more characters. Therefore, the
underscore for information_schema and pg_catalog need to be escaped.

I only realized this when I noticed that :setsp command did not include
the pgstac schema in the search_path! This was a lucky catch.