No fast-forward merge.
Learned from https://nvie.com/posts/a-successful-git-branching-model/
that setting this option allows me to look at what commits were involved
in the feature implementation. The page describes that at the cost of
a few empty commit objects we get the ability to reliably determine the
commits involved in a feature.