Keep the trivia questions around in a session.
Storing the query as a view is not a good idea because invoking the view
will result in a different list every time! I have found that this is
not helpful when i'm still not proficient with what i'm learning. I used
a similar technique in R's repl to improve my proficiency with R (see
`functions_to_learn_this_session` in
https://github.com/vlulla/config/blob/main/Rprofile).