Show function_type also for duckdb's trivia sample
Without the function_type it is easy to forget that the parameter_types
column is listing the arguments that the function takes! For instance
parameter_types says [DOUBLE] for var_samp and what it means is that the
function takes one argument, type DOUBLE, but since var_samp is an
aggregate function it gets an array of double. I was stumped by this
because I tried passing an array of double to var_samp which while
logically correct is not how var_samp is supposed to be used!