Increased the size of history for interactive sessions.
I was looking for some lines which I had used to filter and write forecasts sometime in late october, or early november, and was surprised to learn that I could not find them in my history. I wonder why i had this longer history envvar commented? Anyways, fixing this now!